renamed operation; added error for entryset/sheet name mismatch; expanded readme
