+++ src/main/java/org/opencds/cqf/tooling/OperationFactory.java
@@ -2,7 +2,7 @@ package org.opencds.cqf.tooling;
 
 //import org.opencds.cqf.tooling.jsonschema.SchemaGenerator;
 import org.apache.commons.lang.NotImplementedException;
-import org.opencds.cqf.tooling.terminology.opioidValuesetGenerator.OpioidValueSetGenerator;
+import org.opencds.cqf.tooling.terminology.templateValuesetGenerator.TemplateValuesetGenerator;
 import org.opencds.cqf.tooling.acceleratorkit.DTProcessor;
 import org.opencds.cqf.tooling.acceleratorkit.Processor;
 import org.opencds.cqf.tooling.library.r4.LibraryGenerator;
@@ -14,7 +14,6 @@ import org.opencds.cqf.tooling.qdm.QdmToQiCore;
 import org.opencds.cqf.tooling.quick.QuickPageGenerator;
 import org.opencds.cqf.tooling.terminology.*;
 import org.opencds.cqf.tooling.terminology.distributable.DistributableValueSetGenerator;
-import org.opencds.cqf.tooling.terminology.opioidValuesetGenerator.OpioidValueSetGenerator;
 
 
 class OperationFactory {
@@ -37,8 +36,8 @@ class OperationFactory {
                 return new HEDISValueSetGenerator();
             case "XlsxToValueSet":
                 return new GenericValueSetGenerator();
-            case "OpioidXlsxToValueSet":
-                 return new OpioidValueSetGenerator();
+            case "TemplateValuesetGenerator":
+                 return new TemplateValuesetGenerator();
             case "EnsureExecutableValueSet":
                 return new EnsureExecutableValueSetOperation();
             case "ToJsonValueSetDb":
