+++ src/main/java/org/opencds/cqf/tooling/terminology/templateValuesetGenerator/TemplateValuesetGenerator.java
@@ -1,4 +1,4 @@
-package org.opencds.cqf.tooling.terminology.opioidValuesetGenerator;
+package org.opencds.cqf.tooling.terminology.templateValuesetGenerator;
 
 import ca.uhn.fhir.context.FhirContext;
 import org.apache.poi.ss.usermodel.*;
@@ -14,11 +14,11 @@ import java.io.IOException;
 import java.time.Instant;
 import java.util.*;
 
-public class OpioidValueSetGenerator extends Operation {
+public class TemplateValuesetGenerator extends Operation {
 
     private FhirContext fhirContext;
 
-    public OpioidValueSetGenerator() {
+    public TemplateValuesetGenerator() {
         this.fhirContext = FhirContext.forR4();
     }
 
@@ -34,7 +34,7 @@ public class OpioidValueSetGenerator extends Operation {
         setOutputPath("src/main/resources/org/opencds/cqf/terminology/r4/output");
 
         for (String arg : args) {
-            if (arg.equals("-OpioidXlsxToValueSet")) continue;
+            if (arg.equals("-TemplateValuesetGenerator")) continue;
             String[] flagAndValue = arg.split("=");
             if (flagAndValue.length < 2) {
                 throw new IllegalArgumentException("Invalid argument: " + arg);
@@ -247,31 +247,37 @@ public class OpioidValueSetGenerator extends Operation {
         ValueSet vs;
 
         for (Map.Entry<String, Integer> entrySet : vsMap.entrySet()) {
-            cpgMeta = resolveCpgMeta(workbook.getSheet(entrySet.getKey()));
-            try {
-                if (cpgMeta.getTitle().equals("only fill this out"))
-                    continue;
-            } catch (NullPointerException e) {
-                System.out.println("cpg instance had null title");
-            }
+            Sheet sheet = workbook.getSheet(entrySet.getKey());
+            if (null != sheet) {
+                cpgMeta = resolveCpgMeta(sheet);
+                try {
+                    if (cpgMeta.getTitle().equals("only fill this out"))
+                        continue;
+                } catch (NullPointerException e) {
+                    System.out.println("cpg instance had null title");
+                }
 
-            vs = cpgMeta.populate(fhirContext, outputVersion);
-            meta.populate(vs);
-            resolveCodeList(workbook.getSheet(entrySet.getKey().split("-")[0] + "-cl"), vs, meta.getSnomedVersion());
+                vs = cpgMeta.populate(fhirContext, outputVersion);
+                meta.populate(vs);
+                resolveCodeList(workbook.getSheet(entrySet.getKey().split("-")[0] + "-cl"), vs, meta.getSnomedVersion());
 
 
-            if (vs.hasCompose() && outputVersion.equalsIgnoreCase("r4")) {
-                vs.getMeta().addProfile("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-computablevalueset");
-                vs.addExtension().setUrl("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-knowledgeCapability").setValue(new CodeType("computable"));
-                vs.addExtension().setUrl("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-knowledgeRepresentationLevel").setValue(new CodeType("structured"));
+                if (vs.hasCompose() && outputVersion.equalsIgnoreCase("r4")) {
+                    vs.getMeta().addProfile("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-computablevalueset");
+                    vs.addExtension().setUrl("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-knowledgeCapability").setValue(new CodeType("computable"));
+                    vs.addExtension().setUrl("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-knowledgeRepresentationLevel").setValue(new CodeType("structured"));
+                }
+                if (vs.hasExpansion() && outputVersion.equalsIgnoreCase("r4")) {
+                    vs.getMeta().addProfile("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-executablevalueset");
+                    vs.addExtension().setUrl("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-knowledgeCapability").setValue(new CodeType("executable"));
+                    vs.addExtension().setUrl("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-knowledgeRepresentationLevel").setValue(new CodeType("executable"));
+                    vs.addExtension().setUrl("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-usageWarning").setValue(new StringType("This value set contains a point-in-time expansion enumerating the codes that meet the value set intent. As new versions of the code systems used by the value set are released, the contents of this expansion will need to be updated to incorporate newly defined codes that meet the value set intent. Before, and periodically during production use, the value set expansion contents SHOULD be updated. The value set expansion specifies the timestamp when the expansion was produced, SHOULD contain the parameters used for the expansion, and SHALL contain the codes that are obtained by evaluating the value set definition. If this is ONLY an executable value set, a distributable definition of the value set must be obtained to compute the updated expansion."));
+                }
+                valueSets.add(vs);
             }
-            if (vs.hasExpansion() && outputVersion.equalsIgnoreCase("r4")) {
-                vs.getMeta().addProfile("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-executablevalueset");
-                vs.addExtension().setUrl("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-knowledgeCapability").setValue(new CodeType("executable"));
-                vs.addExtension().setUrl("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-knowledgeRepresentationLevel").setValue(new CodeType("executable"));
-                vs.addExtension().setUrl("http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-usageWarning").setValue(new StringType("This value set contains a point-in-time expansion enumerating the codes that meet the value set intent. As new versions of the code systems used by the value set are released, the contents of this expansion will need to be updated to incorporate newly defined codes that meet the value set intent. Before, and periodically during production use, the value set expansion contents SHOULD be updated. The value set expansion specifies the timestamp when the expansion was produced, SHOULD contain the parameters used for the expansion, and SHALL contain the codes that are obtained by evaluating the value set definition. If this is ONLY an executable value set, a distributable definition of the value set must be obtained to compute the updated expansion."));
+            else{
+                System.out.println("Workbook does NOT contain a sheet named " + entrySet.getKey());
             }
-            valueSets.add(vs);
         }
         return valueSets;
     }
