+++ src/main/java/org/opencds/cqf/tooling/terminology/templateValuesetGenerator/testcase/TestCaseOld.java
@@ -1,4 +1,4 @@
-package org.opencds.cqf.tooling.terminology.opioidValuesetGenerator.testcase;
+package org.opencds.cqf.tooling.terminology.templateValuesetGenerator.testcase;
 
 
 import org.hl7.fhir.DataElement;
